+++ src/main/java/com/edulab/model_base/BaseRole.java
@@ -9,13 +9,13 @@ import com.jfinal.plugin.activerecord.IBean;
 @SuppressWarnings({"serial", "unchecked"})
 public abstract class BaseRole<M extends BaseRole<M>> extends Model<M> implements IBean {
 
-	public M setRoleId(java.lang.Long roleId) {
+	public M setRoleId(java.lang.Integer roleId) {
 		set("role_id", roleId);
 		return (M)this;
 	}
 	
-	public java.lang.Long getRoleId() {
-		return getLong("role_id");
+	public java.lang.Integer getRoleId() {
+		return getInt("role_id");
 	}
 
 	public M setRoleName(java.lang.String roleName) {
